+++ r4/src/test/java/io/github/jy95/fds/r4/translators/BoundsDurationTest.java
@@ -1,8 +1,11 @@
 package io.github.jy95.fds.r4.translators;
 
+import io.github.jy95.fds.common.types.DosageAPI;
 import io.github.jy95.fds.r4.DosageAPIR4;
 import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractBoundsDurationTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Duration;
 import org.hl7.fhir.r4.model.Quantity;
@@ -10,25 +13,16 @@ import org.hl7.fhir.r4.model.Timing;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.ExecutionException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class BoundsDurationTest extends AbstractFhirTest {
+public class BoundsDurationTest extends AbstractBoundsDurationTest<FDSConfigR4, Dosage> {
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoBoundsDuration(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_DURATION);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithBoundsDuration(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateWithBoundsDuration() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Quantity duration = new Duration();
@@ -39,22 +33,24 @@ public class BoundsDurationTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(duration);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_DURATION);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    // For the parametrized test of first test
-    private static String getExpectedText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour 3 jours";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für 3 Tage";
-        } else {
-            return "gedurende 3 dagen";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
+    }
+
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
+    }
+
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
